Migrate Email Connector to JDK 11
