+++ email-ballerina/Ballerina.toml
@@ -15,7 +15,7 @@ version = "@toml.version@"
 "ballerina/io" = "@stdlib.io.version@"
 
 [platform]
-target = "java8"
+target = "java11"
 
     [[platform.libraries]]
     artifactId = "java.mail"
