+++ email-test-utils/build.gradle
@@ -35,3 +35,11 @@ dependencies {
     compile project(":email-native")
 }
 
+compileJava {
+    doFirst {
+        options.compilerArgs = [
+                '--module-path', classpath.asPath,
+        ]
+        classpath = files()
+    }
+}
