+++ email-test-utils/src/main/java/module-info.java
@@ -1,0 +1,9 @@
+module io.ballerina.stdlib.email.testutils {
+    requires org.slf4j;
+    requires greenmail;
+    requires io.ballerina.stdlib.mime;
+    requires java.mail;
+    requires io.ballerina.stdlib.email;
+    requires testng;
+    exports org.ballerinalang.stdlib.email.testutils;
+}
