+++ email-test-utils/src/main/java/org/ballerinalang/stdlib/email/testutils/ImapComplexEmailReceiveTest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.stdlib.email.testutils.mockServerUtils;
+package org.ballerinalang.stdlib.email.testutils;
 
 import com.icegreen.greenmail.user.GreenMailUser;
 import com.icegreen.greenmail.util.GreenMail;
@@ -35,11 +35,11 @@ import javax.mail.internet.MimeMessage;
 import javax.mail.internet.MimeMultipart;
 
 /**
- * Test class for email receive using POP3 with all the parameters.
+ * Test class for email receive using IMAP with all the parameters.
  *
  * @since slp4
  */
-public class PopComplexEmailReceiveTest {
+public class ImapComplexEmailReceiveTest {
 
     private static GreenMailUser user;
     private static final String USER_PASSWORD = "abcdef123";
@@ -63,23 +63,23 @@ public class PopComplexEmailReceiveTest {
     private static final String[] EMAIL_REPLY_TO_ADDRESSES = {"reply1@abc.com", "reply2@abc.com"};
     private static GreenMail mailServer;
 
-    public static Object startComplexPopServer() throws MessagingException {
+    public static Object startComplexImapServer() {
         startServer();
         return null;
     }
 
-    public static Object stopComplexPopServer() {
+    public static Object stopComplexImapServer() {
         mailServer.stop();
         return null;
     }
 
-    public static Object sendEmailComplexPopServer() throws MessagingException {
+    public static Object sendEmailComplexImapServer() throws MessagingException {
         sendEmail();
         return null;
     }
 
     private static void startServer() {
-        mailServer = new GreenMail(ServerSetupTest.POP3);
+        mailServer = new GreenMail(ServerSetupTest.IMAP);
         mailServer.start();
         user = mailServer.setUser(EMAIL_USER_ADDRESS, USER_NAME, USER_PASSWORD);
         mailServer.setUser(EMAIL_USER_ADDRESS, USER_NAME, USER_PASSWORD);
